feat: 라인 밸리데이터 분리
