+++ src/main/java/nextstep/domain/LineValidate.java
@@ -1,0 +1,10 @@
+package nextstep.domain;
+
+import java.util.List;
+
+@FunctionalInterface
+public interface LineValidate {
+
+    boolean validate(List<Boolean> points);
+
+}
